Add integration and unit tests for NNX-based components, including Trainer, Model API, normalizers, encoders, decoders, and couplers. Bug fix in `MLPEncoder`.

+++ src/energnn/trainer/simple_trainer.py
@@ -13,6 +13,7 @@ from typing import Any
 import cloudpickle
 import flatdict
 import jax
+import jax.numpy as jnp
 import numpy as np
 from flax import nnx
 from omegaconf import DictConfig
@@ -273,7 +274,7 @@ class SimpleTrainer:
 
             decision = jax_decision.to_numpy_graph()
             gradient, infos["3_gradient"] = problem_batch.get_gradient(decision=decision, cfg=cfg, get_info=get_info)
-            jax_gradient = JaxGraph.from_numpy_graph(gradient)
+            jax_gradient = JaxGraph.from_numpy_graph(gradient, dtype=jnp.float64)
 
             infos["4_update"] = self.update_params(
                 nnx_optimizer=self.nnx_optimizer,
